HDX-10119 Add a delete_key action to the scan command, only allow resource._csrf_token and extras as deletable keys
